Cygwin changes
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@704 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_callback.c
@@ -641,10 +641,12 @@ mx_invoke_callback( MX_CALLBACK *callback )
 		return MX_SUCCESSFUL_RESULT;
 	}
 
-	MX_DEBUG(-2,("%s: callback->active = %d", fname, callback->active));
+	MX_DEBUG(-2,("%s: callback->active = %d",
+			fname, (int)callback->active));
 
 	if ( callback->active ) {
-		MX_DEBUG(-2,("%s: Aborting since callback->active != 0", fname));
+		MX_DEBUG(-2,
+		("%s: Aborting since callback->active != 0", fname));
 
 		return MX_SUCCESSFUL_RESULT;
 	}
